server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
geojson==3.1.0
shapely==2.0.1
